Do not use method-call syntax on protocol method
GlEnable and GlDisable are protocols. `satisfies?` refers specifically
to protocols. But method-invocation syntax like `.method` only works on
Java interfaces. If the protocol is implemented in-line in a
deftype/defrecord/reify, it will still work, but it will not work with a
pre-existing type that was extended to the protocol.