Move openGL code for selection into scene-editor
There's still a dependency from editors.atlas to
internal.ui.scene-editor, but at least the the structurally-similar
definitions of 'paint-renderer' and 'selection-renderer' are next to one
another.

More refactoring may still be needed around :render-fn and the
exact definition of a "renderable."

Still need to add some kind of broadcast controller to the Atlas Editor
which will forward mouse events to the CameraController and
SelectionController.

Instead of :node-id in the renderable there is now a slightly
more-generic :select-name.

[#86475500]