Modify Selection node on click in Atlas Editor
This has no visible effect in the editor, but you can observe the
"fire-selection-changed" message in the Eclipse console.

I had to move the creation of the Selection node from dynamo.project
into editors.atlas so I could connect it to SelectionController. More
refactoring is probably needed to make this more generic.

Since Selection is no longer created in dynamo.project/make-editor, we
cannot do (.setSelectionProvider site selection-node) in make-editor. Is
this still needed?

[#86475534]