Inject Selection node into SelectionController
The NullPointerException in Selection.getSelection happened because I
was calling .setSelectionProvider on a Selection node inside the same
transaction that was adding it to the graph, so the Selection node I got
was a temporary version that was not yet connected to the world-ref.

This commit reverts the change to dynamo.project/make-editor so that the
selection is constructed there and .setSelectionProvider happens after
the transaction.

Also added inputs and outputs to the SelectionController and Selection
nodes so that they can be connected automatically via injection.

[#86475534]