[WIP] Drag to select in Atlas Editor
Still a bit unsure of the purpose of
com.dynamo.cr.sceneed.ui.RenderView.MIN_SELECTION_BOX in the Java code.
It's set to 16 pixels, which would seem to make it almost impossible to
select small, close together objects. I've set
editors.atlas/min-selection-size to 1 pixel for now, which seems to work
better.

With this commit, click & drag-selection works in the Atlas editor,
including SHIFT/COMMAND for toggling, but it doesn't update the
selection boxes (bright green) while you are dragging the mouse.

[#86475726]