[WIP] Cache get-node-inputs values during drag
get-node-inputs is too expensive to call on every mouse-move. But the
values we need for selection don't change while the a click-and-drag
operation is taking place. So we can cache the values returned by
get-node-inputs in properties on SelectionController.

Now click-and-drag selection is smooth. It would be even better if it
were not calling fire-selection-changed on every mouse-move event: that
still needs fixing.

Click-and-drag combined with multi-select (COMMAND/SHIFT) still needs
work, as it doesn't remember the "previous" selection from before the
click-and-drag started. Maybe we need to add the "pending selection"
concept discussed earlier; or maybe we can cache the previous selection
state in the SelectionController.

[#86475726]