[WIP] Store transient selection state in atom
Following the same approach as in
https://github.com/relevance/defold/commit/a49a45a33f522d961b072d3387cdf1dc79c450b1

We don't want to create a transaction on every mouse-move event during a
click-and-drag, because all those transactions become part of the
undo/redo history.

Instead, we can store some local UI state in an Atom attached as a
property on the SelectionController.

Still to do:

- render the selection box from the state in the atom [#86475828]
- manually trigger a repaint when the mouse moves
- draw selection boxes in response to changes in the atom state
  [#86475856]

[#86475726]