Set *warn-on-reflection* Var root to true
This enables *warn-on-reflection* globally across all Clojure
namespaces.

This is useful for development, but may be more than we want for a
release. It will show reflection warnings even from third-party
libraries which we do not control, such as 'plumbing' and core.cache.

Instead, we could set this Var based on a System Property. Use that for
our own development. For tool developers working within the Defold
editor, we can bind *warn-on-reflection* around the call to
Compiler/load in `internal.clojure`.

[#88045640]