Index arcs by source node.
- Merge all graphs from a transaction into the system, but only record "meaningful" changes in the undo history.
- This excludes the case where a node in a graph is only used as the source of a connection.
- After undo or redo, rebuild all the source (head end) arcs in the changed graph.

In total, this addresses the following case:

When opening a view onto a project node, we need to connect the
project node to the view. We don't want that connection to be subject
to undo. Otherwise, the editor would lose it's updates after the user
hits undo in a totally different tab!