WIP commit.
There is a problem rebuild-sarcs.

One problem was that the :tarcs in the graph being undone were not
being used. (Only the state of the graph _just prior_ to undo was
being used.) I've fixed that.

The remaining problem is with the all-arcs derivation. In the debug
prints, you'll see that :tarcs-from-all-graphs doesn't include the
graph-crossing arc. That's the issue.