Archive cache: Only do atomic file move on same device/disk.
It's important when creating a zip archive that it's created with
one atomic file operation. Otherise we will start serving non-complete
files, which will be corrupt files. This fix ensures that the temporary
file used is created in the same directory as the final zip file in order
to make the atomic rename/move to work. Atomic move is not supported
between two seperate devices/disks.