Override properties invalidate _properties
Setting or clearing a property on an override node (as one does with
instances of scripts in a collection) will cause the node's _properties
output to be invalidated.

The visible behavior of this is in the properties view. Open a
collection, select a script node that is referenced in the
collection. Change a property that was defined in the script code and
observe that the label immediately turns blue on hitting enter or
removing focus from the text field. The reset arrow also appears
immediately. Hitting the reset arrow immediately updated the view as
well.