Node intrinsics are not added if inheriting
Node intrinsics were being injected into every node definition. This
meant that overriding the _properties function would not be inherited,
because a fresh copy of the default function got injected into the
derived node types.

Now, we treat intrinsics as something to inject _only_ when there are no
`inherits` clauses in the node definition. This acts more like a
"universal base type" but without polluting the supertype hierarchy.