Fix hang on open project
Synopsis of the problem: Loading namespaces was tied to displaying the
welcome dialog. On a project open command, the welcome dialog is not
shown, so nothing triggered loading of namespaces. The progress dialog
was waiting for namespaces to finish loading, so it would hang and wait
forever.

Solution: Always trigger loading of namespaces, whether we show the
welcome dialog or not. Previously, main with no args would start loading
namespaces and show the welcome dialog. On completion, the welcome
dialog would call _back_ into main.

Now the welcome dialog uses continuation passing style. `main` always
starts loading namespaces. If no arguments, then show the welcome dialog
with a continuation function that opens the project. (Main is not called
recursively any more.) If there are arguments to main, then call the
function to open the project directly.

In both cases, the new function `open-project-with-progress-dialog`
ensures that namespace loading completes before attempting to use
anything.