Fix nrepl server and middlewares intermittently failing.
After the classloading change in
5e3011a9c59a0581083a0f7b5db153b443ad2784 we've seen nrepl server
intermittently failing to start, and nrepl middlewares (cider/refactor)
consistently failing to load.

Root cause not certain, but seems to be a possible race as nrepl was
started in a separate thread.

This commit starts nrepl in editor boot thread, before requiring boot
namespace, which seems to fix the problem.