Fix problem with properties view in bundle
This change causes the automatically registered value types (Java
classes and Clojure protocols that appear in type position of defnode
clauses) to be carried across the AOT boundary.

We discover these types during macroexpansion and add them to the type
registry. That works fine when running from sources, but when we do AOT,
macroexpansion happens during the compile phase. Once we run from the
resulting uberjar, whatever type information was discovered during
compilation is long gone.

Now we collect the type information during macroexpansion because we
need it later in the code gen process. However, we also emit code from
defnode that will re-register the discovered types at runtime.

This is a tricky process, so I've added some comments in internal.node.