Added INTERPOLATION, TRANSFORM (and more) to jagatoo XMLParam
Those two enums is not implemented by Xith jagatoo and is here added to
support reading of animation (Anims and Bones).
Also added more enums to allow for compatibility above Collada 1.4.0
(this is considered safe as we are not supporting the values as such,
but want to be able to read those files with the values we support).