DEf-1568. Layouts > Defaults randomly gets selected when browsing a texture for a GUI-object node
* The linked folder "builtins", created by IFolder.createLink is
periodically replaced (IResourceDelta.REPLACED) as a link which has the
side-effect of the builtins/../displav profile resource to be reloaded.
This resource should never be changed while the editor is running, so
we skip this event. This behavior has only been reported on Mac OSX,
see DEF-1568. This is a simple (hack) fix for this problem.