DEFEDIT-663 Mitigate some concurrency issues
The graph implementation currently has some issues when the graph is
used concurrently by multiple threads. We avoid some of these issues
by moving the resource-sync! call back on the UI thread. To give a
slightly better UX we instead show a progress dialog informing the
user of what is happening.

Also remove a few cases where we currently interact with the graph
on the UI thread when the UI is disabled. Namely:

- refresh-ui! and refresh-views! are not called when UI is disabled
- events received when UI disabled are explicitly dropped from
  scene when UI is diabled
  - this prevents "Concurrent modification of graph" exceptions when
    some events are delivered while resource sync is running