DEFEDIT-732 Pasting in search dialog also pastes in text editor
Dialogs get their own copy of the main menu with the
system-menu-workaround thingy, a timer for updates etc just like the
main window. We use the contexts of the dialog + global contexts of
the parent scenes for determining what menu items to show, and the
contexts of only the dialog for updating the menu item state (this way
only relevant items will be enabled).