DEFEDIT-859 Fix GUI texture rendering issue
Make sure vertex and UV data line up. For zero-sized GUI nodes (width
and/or height) we did not generate any vertex data, but did generate
UV data. When generating data for a batch with multiple nodes, this
would cause us to, depending on order, use the wrong UVs for some nodes.