No authentication of public endpoints.
This changes the security filter so that it doesn't respond with
"Unauthorized" when trying to access public endpoints. Instead the
user will be authenticated as an anonymous user.

NOTE: A side effect of this is that instead of returning "Unauthorized"
when accessing an non-public endpoint with wrong credentials, it will respond with
"Forbidden" (meaning anonymous user is forbidden to access endpoint).