DEFEDIT-1022 Workararound tree view hangs on Windows 10
This root cause for this issue seems to be similar to an old JavaFX
bug JDK-8132897 which was caused by a combination of having a touch
screen or a tablet attached and clicking an out-of-focus control. That
bug has been fixed but it seems like a similar bug could still be
present for the tree view control we are using.

Workaround is to disable the accessability support in JavaFX by
setting the property "glass.accessible.force" to false.