DEFEDIT-1089 Update embedded lipo to correct executable
macOS ships with a lipo command in `/usr/bin`, which is the one we've
embedded. Unfortunately, this executable isn't actually the lipo
command, just a shim that tries to invoke the real "lipo" through
`libxcselect` ("xcode select") and `xcselect_invoke_xcrun`. This means
that the only reason our bundling has worked is because users have had
xcode installed anyway.

This commit embeds the actual lipo command from the xcode SDK.