DEFEDIT-1209 Fix broken font rendering
* User facing changes

- Text would sometimes be rendered incorrectly when a scene had
  multiple textual objects (labels, gui text nodes) that weren't
  batched together.

* Technical changes

Font rendering uses a glyph cache that maintains a mapping from glyph
to texture coordinates for that glyph within the font texture. As text
is rendered glyphs are looked up or added to the glyph cache, possibly
updating the texture. Previously, due to the lifetime of objects in
the scene cache, the cache would get thrown away but the texture object
would remain, resulting in different text objects having an incorrect
glyph-cache for the active texture object.

This commit removes the "two-stage" glyph-cache + glyph-cell cache,
with a glyph-cache that is just a function from glyph to coordinates,
and this function is shared for all users of the same underlying
texture in a scene.