DEFEDIT-1245 Introduce resource dependencies
We have several places where resource node load functions are
dependent on other nodes having been loaded first. As a workaround
we've either injected load transactions for dependencies in the load
function of the dependent resource, or trusted on the system of
deferred setters and/or in-transaction flag.

Deferred setters run custom setters (and their returned transactions)
as a final recursive step of the transact. This means the setter may
see the result of actions that have happened logically *after* its
position in the transaction list. This happens to solve the resource
dependency problem, but at the cost of being hard to reason
about.

This PR removes deferred setters (runs them where they occur in the
transaction list), the in-transaction flag, and the places where we've
injected dependency load stuff and replaces it all with a way for
resource nodes to declare what other resources it is dependent
on. This information is then used when loading the project and on
resource sync.