DEFEDIT-1427 Properties- and outline panel stop rendering when editing properties while pfx or animation playing
For some reason the rendering of the right split stops entirely if you
run an animation (pfx, sprite) and edit properties and scroll
vigorously in the properties view. We never found the root cause but
it seems to be related to how javafx rendering handles dirty
regions. The ui is not re-rendered until you for instance resize the
panels, but it does react to mouse/keyboard input. The most promising
workarounds were to add a miniscule padding to the scroll-bar arrows,
or to avoid nesting the ScrollPane in a TitledPane. We choose the
latter. Since we weren't using any of the TitledPane's fancy features
anyhow, we replaced it with a plain VBox + Label.

* User facing changes
  - No more hangs in the properties view while playing pfx or
  animations

* Technical changes
  - Scene graph changes in .fxml + styling to look like a TitledPane.
  - Unrelated - removed small padding around ScrollPane, visible as a
  line under the label.