DEFEDIT-1494 IllegalArgumentException after zooming in/out
This happens if you scroll with an apple mouse and close the tab while
the inertia is still keeping the scroll going. Execeptions in scene &
code view. Don't fully understand why javafx chooses to dispatch
events to nodes that should no longer be part of the scene
hierarchy (closed tab). This workaround sets a custom "null" event
dispatcher for the views when the tab closes. This way we don't have
to track every single event listener and unregister manually.