Make shader binding & scene cache update more stable.
When running the editor in dev mode any shader syntax error throws a
glGetError related exception because we attempt to delete an invalid
program number. During scene cache update the old program is first
deleted, then we try to create a new one but fail to consider
exceptions properly. Now we fall back to an invalid "0" program with
no uniforms.