Issue 4311 sys.get ifaddrs() (#4373)
* Also include AF_INET6

* Add more than one address per interface

* Add "family" to return value from sys.get_ifaddrs()

* Update socket_win32.cpp

Iterate over all addresses of an interface. Distinguish between ipv4 and ipv6 and link.

* m_Family -> m_family

* Fixed failing SSDP test

We can now get both ipv4 and ipv6 addresses for a single network interface on macOS. Previously we only got one ipv4. The test always expected the socket bound to an address to be valid, but ssdp skips ipv6 addresses which caused the test to fail with the changes in this PR.

* Set the physical address (MAC) on all returned results