Issue-5843 Disable component types that are not used in a collection (#6062)
Bob analyse collections to detect what component types that are used and the engine only allocates memory for the collections that are used. And if the collection has not factories we can know exactly how many components that are used. But components in collections we use max_count from `game.project`