Set layout node names from display profiles
Note: the previous implementation of the `add-handler-info` label
returned a single tuple with information about a single context menu
option. Gui's LayoutsNode has many add options — one for every not yet
defined display profile. I cannot change it to always return
a collection of tuples, because this is label is also used in spine
extension, and since we can't force our users to update, such a change
will break existing users. So I decided to allow both a single tuple and
a collection of tuples to preserve compatibility.

Note: there is a `handler-infos` input label that collects all handler
info tuples from GuiSceneNode's children. Due to a data type change of
`add-handler-info` I also added a `handler-infos` output that ensures
that it's a flat vector of info tuples. My concern is: `:handler-infos`
are completely unused. Should we remove them altogether?