Fix move-up/move-down action availability
The problem with existing availability implementation was the assumption
that `:child-indices` of a parent are consequtive and start with 0. It
turns out `:child-indices` are neither. For nested gui nodes, child
indices don't start with 0, so it's possible to have indices in the
following shape:
```clj
[[72057594037928031 4]
 [72057594037928032 5]]
```
Also, after adding new box nodes, it's possible for indices to be
non-consequtive, e.g.:
```clj
[[72057594037928029 2]
 [72057594037928030 3]
 [72057594037928033 6]]
```
The solution is to find min/max index and compare it with selected
node's index.

Solves #5811