Preserve changes-view on failed "git status"
When `git status` fails during project startup, `make-changes-view` will
return `nil`. This makes it throw NPEs on creating and deleting files,
but not during resource sync that check if for changes-view nullability.
I think resource sync does it only so it's easier to test it, since
other places in code expect `changes-view` to exist. With these changes,
exceptions during `git status` execution will be shown to the user on
resource sync, creation and deletion of files.

I think this behavior will be better than the current one because
the shown exception is more relevant to the user. For example, in the
related issue the exception looks like that:
```
Malformed input or input contains unmappable characters: /home/pjarosz4/Desktop/witchcrafter/assets/images/gui/backpack_bg ��� kopia.png
```
This error can be solved by the user.

Fixes #6623