Support editing empty form fields (#6973)
Usually we provide default values to form input views if they are absent in the form data, but it's not possible to do in some contexts like nils in map vals in vectors edited with `:table` view. This commit ensures nil-sensitive form input views work as expected when they receive nil values.