Add support for showing diagnostics in the editor
This is a first changeset in series of adding linting support to the editor. It adds diagnostics as a concept and renders them in the code view. Diagnostics are regions (aka cursor ranges) with some extra fields that will be published by language servers to the editor and shown in code view. This changeset allows showing diagnostics, but does not yet have any support for actually creating diagnostics. To test it, you can do the following:

1. Open code view and select some text (multi-selection works too)
2. Evaluate this in the REPL:

       (let [code-view (-> 0
                           (dynamo.graph/targets-of :build-settings)
                           ffirst
                           (dynamo.graph/targets-of :_node-id)
                           ffirst
                           (dynamo.graph/node-value :active-view))
             ranges (dynamo.graph/node-value code-view :cursor-ranges)
             resource (dynamo.graph/node-value (dynamo.graph/node-value code-view :resource-node) :resource)
             diagnostics (mapv #(let [s (rand-nth [:error :warning :information :hint])]
                                  (assoc % :type :diagnostic
                                           :severity s
                                           :message (str "Severity is " s)))
                               ranges)]
            (:status (dynamo.graph/transact
                       (dynamo.graph/update-property 0 :resource->diagnostics assoc resource diagnostics))))

I don't plan to merge this to `editor-dev` yet. Instead, I'll merge other LSP-related changesets to this branch and, when everything is done, I will finally merge it to `editor-dev`.

Related to #3885