Improve binary plugins support (#7170)
This changeset instructs the editor and bob to skip the engine build if an extension does not have a source folder (`src` or `commonsrc`). This allows us to use native extensions' unpacking and plugins without having to define an engine extension. Additionally, this changeset marks unpacked files from the `plugins/bin` folder as executable.

Related to #3885