Add a dev tool to find dangling graph outputs
This tool helps to find outputs that are not connected to anything and potentially remove unnecessary outputs. Unfortunately, it seems we don't have dangling outputs that contribute to successors of highly referenced nodes. Besides `sha256`, that is. But it seems we need it in the graph for caching. Regardless, it highlights some outputs that we might want to remove in the future to not clutter the code (e.g. `breakpoint-rows`, `gui-scene`).

Related to #5447