Identical names require the same values in templates
User-facing changes: when multiple templates used in the same GUI define resources with clashing names, we now require them to refer to the same values. For instance, suppose a GUI file named `A.gui` contains two templates: `B.gui` and `C.gui`. If both templates define a font named "main_font" and the actual fonts used in each template are not the same, `A.gui` will not know which one to use and will randomly choose one. To prevent this issue, we will now display a build error when such clashes occur.

Fixes #7222