Add an option to open a project from the editor (#7539)
User-facing changes: Added a `Open Project...` menu item in the File menu. It will launch another instance of the editor.

Technical notes: The code to launch the process is based on the code in the `updater.clj` file. A notable modification has been made to the file unpacking process. Instead of using a stable folder in the "editor support path" location, we now always unpack files to a temporary folder. It turned out that using the support path was a mistake because it is intended for files that persist between application and system restarts, whereas the unpacked files were always deleted when the editor was closed. This meant that the behavior of the unpacked files was similar to that of a temporary directory, but they were being stored in the wrong place. This shared storage location was causing issues with running multiple instances of the editor at the same time — shutdown of one instance of the editor broke another.

Fixes #5689