Quick fix for issue-7531
The ES3 shader transform doesn't seem to handle "non-directive" lines when a fragment shader starts with a comment. This only happens in the editor, so for now we simply strip all the comments before doing the transform. Bob already deals with stripped source code from the shader preprocessor, so we might want to take a look at how we want to do this a bit later..