Add override inspector
User-facing changes:
This changeset adds an override inspector view that can be requested using the Show Overrides context menu action. This action is available for items in the Outline View and for individual properties in the Properties View. When the action is executed, the editor will search through the uses of this item and show every place that changes the properties of the item. In this override inspector table, you can double-click on a result row to open the place where the properties are overridden.

Fixes #7127