Fix focus on committing values in the form view
Technical notes:
When editing fields in the `game.project` it was possible to lose the input focus after committing the value if the value became clearable. This happened because a new node — reset button — was added to the grid pane children, and it caused the re-creation of this and subsequent inputs, so the view lost focus. Using `:fx/key` stabilized the component description -> JavaFX node mapping, enforcing that the input stays the same object and is not re-created.

User-facing changes:
We don't lose input focus when editing `game.project` fields.

Fixes #7613