Show notifications on resource collisions
User-facing changes:
When a local folder overrides a zip folder from a dependency, we now show a warning notification in the editor.

Technical notes:
This changeset implements a notification system in the editor. Resource sync diffs its errors with the old snapshot to show/hide notifications about folder overrides. I initially considered adding a "Rename Folder" action to the notification, but it required moving the renaming logic from the `editor.asset-browser` to the `editor.workspace` namespace that is UI-dependent (e.g. it shows a rename dialog, a resolve conflicts dialog). While it's doable, I don't like that it makes the workspace ns even more dependent on UI than it currently is. I think it should have logic only so we can eventually use it in Bob, but it will then require either adding a layer of indirection for "non-GUI UI". Another solution would be moving the collision notification handling somewhere out of `editor.workspace` ns (e.g. an error listener). Not showing any actions is a also viable solution here, since the notification is useful as it is, we can deliver stuff earlier to our users, and we still can add actions later if our users will need them.

Fixes #7375