Check node existence before querying it on load
Technical notes:
The problem is that `editor.defold-project/connect-resource-node` might return a node id for a node that does not yet exist in the system in case where it has to create a new node. We cannot query such nodes using `g/node-value` because they don't exist yet. Users of `connect-resource-node` first need to check that the node already exists if they want to use `g/node-value` on it.

User-facing changes:
Projects that previously couldn't be opened due to a broken resource references in files now can be opened.

Fixes #7174