Support multiple materials in models
User-facing changes:
Models will migrate to the new syntax that supports multiple materials — you will need to open the project in the editor and re-save it to use the new syntax. You should update the material list by opening the models, copying data from the "default" material to the actual material used by the gltf/collada file, and then removing the "default" material, although it will continue working as.

Technical notes:
The multi-material support is implemented using Material node overrides that set the texture bindings. This enables automatic rename of texture bindings in the model file when changing sampler names in the material file.

Fixes #8079
Fixes #7722