Support multiple materials in models (#8186)
* Support multiple materials in models

User-facing changes:
Models will migrate to the new syntax that supports multiple materials — you will need to open the project in the editor and re-save it to use the new syntax. You should update the material list by opening the models, copying data from the "default" material to the actual material used by the gltf/collada file, and then removing the "default" material, although it will continue working as.

Technical notes:
The multi-material support is implemented using Material node overrides that set the texture bindings. This enables automatic rename of texture bindings in the model file when changing sampler names in the material file.

Fixes #8079
Fixes #7722

* Review fixes

* Review fixes

* Re-implement models without overrides

* Test fixes

* Add newline

* Add deletion inference

* Fix test data

* Use black pixel texture when there are no textures

* Review fixes

* Refactor name-index detectors

* Precompute sampler name change-related info

* Mimic engine behavior w.r.t. texture selection

* Include all meshes when calculating AABB

* Resaved hand-edited test data in test project

---------

Co-authored-by: Mats Gisselson <mats.gisselson@gmail.com>