Wait for the reload to finish before building
When the user focuses on the editor, it performs a resource reload. On large projects this may take a while, and if the user starts a build, the build will use outdated resources. This change fixes the issue by waiting for the reload to finish before starting the build.

Fixes #6681