Edit collision shapes using editor scripts
It is now possible to edit collision shapes using editor scripts! To support this, we added new `"shapes"` node list property to collision objects. For example, adding a new shape is done like this:
```lua
editor.transact({
    editor.tx.add("/hero.collisionobject", "shapes", {
        type = "shape-type-box" -- or "shape-type-sphere", "shape-type-capsule"
    })
})
```

To create a new collision shape, it is necessary to specify the type of the shape. The available types are:
- `shape-type-box` - box shape with `dimensions` property
- `shape-type-sphere` - sphere shape with `diameter` property
- `shape-type-capsule` - capsule shape with `diameter` and `height` properties

Related to #8504