Don't use `.` and `(` as completion auto-commit characters (#10888)
Previously, the text editor would accept displayed completion if the user typed `.` or `(`. It results in a surpising behavior if the user was simply typing and didn't look at the completion list. Other editors like vscode don't accept these characters as completion auto-commit characters, so we follow the same behavior.

Related to #10781